# fix(client): Fix batch deletes