--- a/.changeset/fast-poems-perform.md
+++ b/.changeset/fast-poems-perform.md
@@ -1,0 +1,5 @@
+---
+"electric-sql": patch
+---
+
+Fix `DELETE` statement batching causing call stack overflow and improve performance.