--- a/clients/typescript/test/migrators/builder.ts
+++ b/clients/typescript/test/migrators/builder.ts
@@ -417,14 +417,20 @@ export const builderTests = (test: TestFn<ContextType>) => {
         ? ['?', '?', '?', '?']
         : ['$1', '$2', '$3', '$4']
 
+    const withValues = builder.dialect !== 'SQLite'
+
     t.deepEqual(stmts, [
       {
-        sql: `DELETE FROM test WHERE ("a" = ${posArgs[0]} AND "b" = ${posArgs[1]}) OR ("a" = ${posArgs[2]} AND "b" = ${posArgs[3]})`,
+        sql: `DELETE FROM test WHERE ("a", "b") IN (${
+          withValues ? ' VALUES ' : ''
+        }(${posArgs[0]}, ${posArgs[1]}), (${posArgs[2]}, ${posArgs[3]}))`,
 
         args: [1, 2, 3, 4],
       },
       {
-        sql: `DELETE FROM test WHERE ("a" = ${posArgs[0]} AND "b" = ${posArgs[1]})`,
+        sql: `DELETE FROM test WHERE ("a", "b") IN (${
+          withValues ? ' VALUES ' : ''
+        }(${posArgs[0]}, ${posArgs[1]}))`,
         args: [5, 6],
       },
     ])
@@ -449,13 +455,19 @@ export const builderTests = (test: TestFn<ContextType>) => {
         ? ['?', '?', '?', '?']
         : ['$1', '$2', '$3', '$4']
 
+    const withValues = builder.dialect !== 'SQLite'
+
     t.deepEqual(stmts, [
       {
-        sql: `DELETE FROM test WHERE ("b" = ${posArgs[0]} AND "a" = ${posArgs[1]}) OR ("b" = ${posArgs[2]} AND "a" = ${posArgs[3]})`,
+        sql: `DELETE FROM test WHERE ("b", "a") IN (${
+          withValues ? ' VALUES ' : ''
+        }(${posArgs[0]}, ${posArgs[1]}), (${posArgs[2]}, ${posArgs[3]}))`,
         args: [2, 1, 4, 3],
       },
       {
-        sql: `DELETE FROM test WHERE ("b" = ${posArgs[0]} AND "a" = ${posArgs[1]})`,
+        sql: `DELETE FROM test WHERE ("b", "a") IN (${
+          withValues ? ' VALUES ' : ''
+        }(${posArgs[0]}, ${posArgs[1]}))`,
         args: [6, 5],
       },
     ])